worker/requirements.txt
aiomysql==0.2.0
redis==5.0.1
apscheduler==3.10.4
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
structlog==23.2.0
python-json-logger==2.0.7
PyYAML==6.0.1
